move request permissions in mainactivity, in android 6 they cant be launched outside from an activity
